feature/REG-74 rename DidCryptoSuit.ts
