+++ packages/did-hub-client/src/crypto/index.ts
@@ -1,4 +1,4 @@
-export * from './DidCryptoSuit';
+export * from './DidCryptoSuite';
 export * from './DidPrivateKey';
 export * from './DidPublicKey';
 export * from './PublicKey';
