+++ packages/did-hub-client/src/utils.ts
